implemented Hotkey logic inside Options
Only the comes and goes between the new Hotkeys section in Options and the new "Hotkeys.xml" file has been implemented.
Whatever Hotkey is specified in that window, will get saved within a second of being input, to the XML file. Every time the Options window is opened, the Hotkeys list gets filled with what is read from the XML.

I still need to implement the logic that will allow the usage of those custom hotkeys to actually execute commands inside the editor.