fixed selecting compile errors not showing in files
This is an initial approach to fixing the editor not going to the lines that have errors when they are clicked in the Error Grid when compiling.

This fix aims to work at plugins compiled against SM 1.11 and later - previous versions don't show errors that happened in external files to the main file properly, and therefore the interactive feature in the editor will obviously not work.

If a subfile has an error, the current approach goes about looping through all compiled scripts, searching through their dirs looking for the subfile with the error.