fixed recent files list filling up
Previously, the Recent Files menu item was getting filled up with all of the files that were previously opened in the editor. Now it adds them to the list only if the editor is fully initialized, which means it will not add new files if they're being loaded on startup.