fixed program crashing when compiling (read desc)
In this commit, I changed the way the program handles compilation errors.

Previously, if a plugin failed to compile, the program would redirect the errors to a text file, and read them off of there. I believe certain crashes were related to the program trying to read the file at the same time as some other file sync program (like Dropbox, Drive, etc.) while working inside synced directories.

Now, the editor reads the compiler process standard output and extracts errors with a Regex in case the compilation was unsuccessful, avoiding to rely on external files.